fix: :bug: bug fixes
Bug fixes are being made, styles are being added to the page, certain functionalities are still missing
