+++ css/style-student.css
@@ -187,15 +187,26 @@ main{
     padding-top: 60px;
 }
 .div__header{
-    background-color: #f9f9f9;
+
     padding: 10px 0;
     display: flex;
     justify-content: space-between;
-    padding: 25px;
+    align-items: center;
+    padding: 15px;
 }
 .div__header h2 {
     font-size: 22px;
     font-family: "Montserrat-Bold";
+}
+.div__header div {
+    display: flex;
+    border: none;
+    padding-right:30px ;
+}
+.sort__button{
+    border: none;
+    padding: 5px 10px;
+    background: transparent;
 
 }
 .button__add {
@@ -231,6 +242,7 @@ main{
     font-family: "Montserrat-Bold";
     text-align: center;
 }
+
 #dialog__close,
 #dialog__close_edit{
     width: 30px;
@@ -239,6 +251,12 @@ main{
     border-radius: 10px;
     display: flex;
     justify-content: center;
+}
+.dialog__main strong{
+    font-size: 15px;
+    font-family: "Montserrat-Medium";
+    color: var(--color-6);
+
 }
 form input{
     width: 100%;
@@ -329,6 +347,12 @@ th {
         min-height: 105dvh;
 
     }
+    #img__back{
+        width: 18px;
+        height: 18px;
+        margin-left: 170px;
+        margin-right: 20px;
+    }
     .div__search{
         max-width: 152px;
         width: 100%;
@@ -343,12 +367,37 @@ th {
         width: 100%;
         gap: 15px;
         padding-top: 20px;
+        display: flex;
+        grid-area: 1 / 1 / 6 / 4;
+        padding-left: 160px;
+        padding-top: 60px;
+    }
+    .div__header{
+        gap: 10px;
+        padding: 10px;
+        padding-left: 20px;
+        display: flex;
+        align-items: center;
+        
+    }
+    .div__header h2 {
+        font-size: 12px;
+    }
+    .div__header div {
+        display: flex;
+        border: none;
+        padding-right:10px ;
+    }
+
+    .button__add {
+        padding: 5px 15px;
+        font-family: "Montserrat-Medium";
+        font-size: 9px;
     }
     .Section__header{
         padding-left: 5px;
         padding-right: 5px;
         width: 98%;
-
     }
     .left__nav {
         max-width: 150px;
