+++ js/student.js
@@ -1,6 +1,5 @@
-import {save, edit, transformInputFormulary, loadStudent, showRowsTable} from "./components/formStudent.js";
+import {save,edit, transformInputFormulary, loadStudent, showRowsTable} from "./formStudent.js";
 
-// Creacion de variables con selector al HTML por ID
 const add__student = document.querySelector("#add__student");
 const dialog__add_student = document.querySelector("#dialog__add_student");
 const dialog__student_edit = document.querySelector("#dialog__student_edit");
@@ -10,19 +9,20 @@ const dialog__student_delete = document.querySelector("#dialog__student_delete")
 const form__student = document.querySelector("#form__student");
 const form__student_edit = document.querySelector("#form__student_edit");
 const form__student_delete = document.querySelector("#form__student_delete");
-const table__student = document.querySelector("#table__student");
+const table__student = document.querySelector("#table__student")
 
-// showRowsTable(loadStudent());
+
 addEventListener("DOMContentLoaded", (e)=>{
     const DB = loadStudent();
     showRowsTable(DB);
 })
 
+
 add__student.addEventListener("click", ()=>{
     dialog__add_student.showModal();
-}) //Crear evento de escucha para abrir cuadro de dialogo para adicionar un estudiante
+})
 
-dialog__close.addEventListener("click", ()=> dialog__add_student.close()); //Crear evento de escucha para cerrar el cuadro de adicionar un estudiante 
+dialog__close.addEventListener("click", ()=> dialog__add_student.close());
 dialog__close_edit.addEventListener("click", ()=> dialog__student_edit.close());
 
 form__student.addEventListener("submit", (e)=>{
@@ -38,7 +38,9 @@ form__student_edit.addEventListener("submit", (e)=>{
     const response = edit(data);
     console.log(response);
     form__student_edit.reset();
+    
 })
+
 form__student_delete.addEventListener("submit", (e)=>{
     e.preventDefault();
     const actions = e.submitter.dataset.action; // delete o close
